pyproject.toml
[tool.poetry]
name = "unxpass"
version = "0.1a"
description = "Measuring soccer player's creativity"
authors = ["Pieter Robberechts <pieter.robberechts@kuleuven.be>"]
readme = "README.md"
homepage = "https://github.com/ML-KULeuven/un-xPass"
repository = "https://github.com/ML-KULeuven/un-xPass"
classifiers = [
  "Programming Language :: Python :: 3",
  "License :: OSI Approved :: Apache Software License",
  "Operating System :: OS Independent"
]

[tool.poetry.urls]
Changelog = "https://github.com/ML-KULeuven/un-xPass/releases"

[tool.poetry.dependencies]
python = ">=3.9,<4.0"
rich = "^13.4.2"
typer = "^0.9.0"
mlflow = "^2.4.2"
xgboost = "^1.7.6"
scikit-learn = "^1.3.0"
torch = "1.13.1"
lightning = "^2.0.5"
socceraction = "^1.4.1"
mplsoccer = "^1.1.12"
statsbombpy = "^1.10.1"
pyarrow = "^12.0.1"
hydra-core = "^1.3.2"
hydra-optuna-sweeper = { version = "^1.2.0", optional = true }
hydra-ray-launcher = { version = "^1.2.0", optional = true }
tables = { version = "^3.8.0", optional = true }

[tool.poetry.extras]
experiments = ["hydra-optuna-sweeper", "hydra-ray-launcher"]
hdf = ["tables"]

[tool.poetry.scripts]
unxpass = "unxpass.__main__:app"

[tool.poetry.group.dev.dependencies]
black = "^23.7.0"
isort = "^5.12.0"
flake8 = "^6.0.0"
mypy = "^1.4.1"
pre-commit = "^3.3.3"
pre-commit-hooks = "^4.4.0"
darglint = "^1.8.1"
pep8-naming = "^0.13.3"
flake8-docstrings = "^1.7.0"
flake8-rst-docstrings = "^0.3.0"
pyupgrade = "^3.9.0"

[tool.poetry.group.test.dependencies]
pytest = "^7.4.0"
pytest-lazy-fixture = "^0.6.3"

[tool.black]
line-length = 99
target-version = ['py39']
include = '\.pyi?$'

[tool.isort]
profile = "black"
force_single_line = false
lines_after_imports = 2

[tool.mypy]
strict = true
warn_unreachable = true
pretty = true
show_column_numbers = true
show_error_codes = true
show_error_context = true
ignore_missing_imports = true

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
